requirements.txt
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
lightgbm==4.6.0
xgboost==3.0.5
statsmodels==0.14.5
prophet==1.1.7
tensorflow==2.19.0
matplotlib==3.10.0
tqdm==4.67.1
pyyaml==6.0.2
